requirements.txt
# Presentation and PDF processing
python-pptx>=0.6.21
pdf2image>=1.16.3
Pillow>=10.0.0

# Font extraction and hunting
fonttools>=4.43.0
requests>=2.31.0

# Environment variables
python-dotenv>=1.0.0

# Keynote file processing (requires macOS)
# pyobjc-framework-Cocoa>=9.0 (for macOS Keynote support)

# File handling
zipfile36>=0.1.3

# CLI interface
click>=8.1.7
colorama>=0.4.6
tqdm>=4.66.1

# PDF generation (for reports)
reportlab>=4.0.0

# Web interface
Flask>=3.0.0
Werkzeug>=3.0.0

